Remove unnecessary files, Created folder structure
