+++ src/App.js
@@ -1,24 +1,20 @@
-import logo from './logo.svg';
-import './App.css';
+import React from "react";
+import {BrowserRouter as Router, Switch, Route} from 'react-router-dom';
+import Inventory from "./Pages/Dashboard/Inventory";
 
-function App() {
+import Home from "./Pages/Home/Home";
+
+
+
+const App = () => {
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
+  <Router>
+   <Switch>
+     <Route path="/" exact component={Home}/>
+     <Route path="/inventory" exact component={Inventory}/>
+     <Route exact path='*'component={Home}/> 
+   </Switch>
+   </Router>
   );
 }
