+++ src/index.css
@@ -5,7 +5,12 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
+  background-color: #181818;
+  color: #fff;
+  
 }
+
+/* #222222 */
 
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
